// backend/frontend/packages/promtube-backend/models/GetUnapprovedVideos200ResponseInner.ts
/**
 * Promtube Backend API
 * The API for Promtube
 *
 * OpenAPI spec version: 0.0.1
 * Contact: horahora1567@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class GetUnapprovedVideos200ResponseInner {
    'videoID'?: number;
    'URL'?: string;
    'category'?: number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "videoID",
            "baseName": "VideoID",
            "type": "number",
            "format": ""
        },
        {
            "name": "URL",
            "baseName": "URL",
            "type": "string",
            "format": ""
        },
        {
            "name": "category",
            "baseName": "Category",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return GetUnapprovedVideos200ResponseInner.attributeTypeMap;
    }

    public constructor() {
    }
}