// backend/frontend/packages/promtube-backend/models/Videos200ResponseVideosInner.ts
/**
 * Promtube Backend API
 * The API for Promtube
 *
 * OpenAPI spec version: 0.0.1
 * Contact: horahora1567@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class Videos200ResponseVideosInner {
    'title'?: string;
    'videoID'?: number;
    'views'?: number;
    'authorID'?: number;
    'authorName'?: string;
    'thumbnailLoc'?: string;
    'rating'?: number;
    'videoDuration'?: number;
    'isMature'?: boolean;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "title",
            "baseName": "Title",
            "type": "string",
            "format": ""
        },
        {
            "name": "videoID",
            "baseName": "VideoID",
            "type": "number",
            "format": ""
        },
        {
            "name": "views",
            "baseName": "Views",
            "type": "number",
            "format": ""
        },
        {
            "name": "authorID",
            "baseName": "AuthorID",
            "type": "number",
            "format": ""
        },
        {
            "name": "authorName",
            "baseName": "AuthorName",
            "type": "string",
            "format": ""
        },
        {
            "name": "thumbnailLoc",
            "baseName": "ThumbnailLoc",
            "type": "string",
            "format": ""
        },
        {
            "name": "rating",
            "baseName": "Rating",
            "type": "number",
            "format": ""
        },
        {
            "name": "videoDuration",
            "baseName": "VideoDuration",
            "type": "number",
            "format": ""
        },
        {
            "name": "isMature",
            "baseName": "IsMature",
            "type": "boolean",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Videos200ResponseVideosInner.attributeTypeMap;
    }

    public constructor() {
    }
}